readme and code rearrange done
