+++ pom.xml
@@ -92,26 +92,6 @@
 		</dependency>
 	</dependencies>
 
-    <profiles>
-        <profile>
-            <id>local</id>
-            <properties>
-                <logstash.port>10000</logstash.port>
-                <logger.console.tresholdFilter>ALL</logger.console.tresholdFilter>
-                <logger.socket.tresholdFilter>OFF</logger.socket.tresholdFilter>
-            </properties>
-        </profile>
-
-        <profile>
-            <id>dev</id>
-            <properties>
-                <logstash.port>10001</logstash.port>
-                <logger.console.tresholdFilter>OFF</logger.console.tresholdFilter>
-                <logger.socket.tresholdFilter>ALL</logger.socket.tresholdFilter>
-            </properties>
-        </profile>
-    </profiles>
-
 	<build>
 		<plugins>
 			<plugin>
