+++ src/main/java/pl/iordervivi/data/scrapper/gov/covid19/job/DataForCovid19StatisticWebScrapperJob.java
@@ -6,8 +6,9 @@ import org.quartz.JobExecutionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
-import pl.iordervivi.data.scrapper.gov.covid19.service.DataProviderImpl;
+import pl.iordervivi.data.scrapper.gov.covid19.service.DataProvider;
 
 import javax.transaction.Transactional;
 
@@ -16,9 +17,11 @@ import javax.transaction.Transactional;
 public class DataForCovid19StatisticWebScrapperJob implements Job {
     private final static String jobStarted = "Data web scrapping for covid-19 disease starting";
     private static Logger logger = LoggerFactory.getLogger(DataForCovid19StatisticWebScrapperJob.class);
+
     //TODO: fix in future with static constructor and interface throw error
     @Autowired
-    private DataProviderImpl dataProvider;
+    @Qualifier("DataProviderImpl")
+    private DataProvider dataProvider;
 
     @Override
     public void execute(JobExecutionContext jobExecutionContext) throws JobExecutionException {
