+++ src/main/java/pl/iordervivi/data/scrapper/gov/covid19/repo/RegionRepository.java
@@ -11,6 +11,5 @@ import java.util.Optional;
 @Repository
 public interface RegionRepository extends JpaRepository<Region, Integer> {
     Optional<Region> findByRegion(String region);
-
     boolean existsByRegion(String region);
 }
