+++ src/main/java/pl/iordervivi/data/scrapper/gov/covid19/service/DataProvider.java
@@ -1,5 +1,5 @@
 package pl.iordervivi.data.scrapper.gov.covid19.service;
 
 public interface DataProvider {
-    boolean addSicksByRegion();
+    void addSicksByRegion();
 }
